return informations when executing activities and transitions
git-svn-id: http://svn.egroupware.org/egroupware/trunk@19098 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ workflow/inc/engine/src/API/Instance.php
@@ -149,10 +149,10 @@ class Instance extends Base {
   }
  
   /*!
-   \private
-   Creates a new instance.
-   This method is called in start activities when the activity is completed
-   to create a new instance representing the started process.
+  * @private
+  * Creates a new instance.
+  * This method is called in start activities when the activity is completed
+  * to create a new instance representing the started process.
   */
   function _createNewInstance($activityId,$user) {
     // Creates a new instance setting up started,ended,user
@@ -351,7 +351,8 @@ class Instance extends Base {
   * concurrent access to this function is normally handled by WfRuntime and WfSecurity
   * theses objects are the only ones which should call this function. WfRuntime is handling the
   * current transaction and WfSecurity is Locking the instance and instance_activities table on
-  * a 'run' action which is the action leading to this setActivityUser call
+  * a 'run' action which is the action leading to this setActivityUser call (could be a release 
+  * as well on auto-release)
   * @param $activityId is the activity Id
   * @param $theuser is the user id or '*' (or 0, '' or null which will be set to '*')
   * @return false if something was wrong
@@ -365,11 +366,11 @@ class Instance extends Base {
         $found = true;
 
         // prepare queries
-        $where = " where `wf_activity_id`=? and `wf_instance_id`=?";
+        $where = ' where wf_activity_id=? and wf_instance_id=?';
         $bindvars = array((int)$activityId,(int)$this->instanceId);
         if(!($theuser=='*')) 
         {
-          $where .= " and (`wf_user`=? or `wf_user`=?)";
+          $where .= ' and (wf_user=? or wf_user=?)';
           $bindvars[]= $theuser;
           $bindvars[]= '*';
         }
@@ -378,6 +379,7 @@ class Instance extends Base {
         $query = 'update '.GALAXIA_TABLE_PREFIX.'instance_activities set wf_user=?';
         $query .= $where;
         $bindvars_update = array_merge(array($theuser),$bindvars);
+  echo "<br>DEBUG:".$query; _debug_array($bindvars_update);
         $this->query($query,$bindvars_update);
         $this->activities[$i]['wf_user']=$theuser;
         return true;
@@ -506,15 +508,15 @@ class Instance extends Base {
   
   //! Completes an activity, normally from any activity you should call this function without arguments.
   /*!
-  The arguments are explained just in case.
-  $activityId is the activity that is being completed, when this is not
-  passed the engine takes it from the $_REQUEST array,all activities
-  are executed passing the activityId in the URI.
-  $addworkitem indicates if a workitem should be added for the completed
-  activity.
-  YOU MUST NOT CALL complete() for non-interactive activities since
-  the engine does automatically complete automatic activities after
-  executing them.
+  * YOU MUST NOT CALL complete() for non-interactive activities since
+  * the engine does automatically complete automatic activities after
+  * executing them.
+  * The arguments are explained just in case. YOU SHOULD NOT USE THEM in your activity code.
+  * @param $activityId is the activity that is being completed, when this is not
+  * passed the engine takes it from the $_REQUEST array,all activities
+  * are executed passing the activityId in the URI.
+  * @param $addworkitem indicates if a workitem should be added for the completed
+  * activity (true by default).
   */
   function complete($activityId=0,$addworkitem=true) {
     global $user;
@@ -567,81 +569,99 @@ class Instance extends Base {
     $this->setActivityStatus($activityId,'completed');
     
     //If this and end actt then terminate the instance
-    if($type=='end') {
-      $this->terminate();
-      return;
+    if($type=='end') 
+    {
+      return $this->terminate();
+    }
+    else
+    {
+      return true;
     }
     
   }
   //! Send autorouted activities to the next one(s). Private engine function
   /*
-  The arguments are explained just in case.
-  $activityId is the activity that is being completed, when this is not
-  passed the engine takes it from the $_REQUEST array,all activities
-  are executed passing the activityId in the URI.
-  $force indicates that the instance must be routed no matter if the
-  activity is auto-routing or not. This is used when "sending" an
-  instance from a non-auto-routed activity to the next activity.
-  YOU MUST NOT CALL sendAutorouted() for non-interactive activities since
-  the engine does automatically complete and send automatic activities after
-  executing them.
-  This function is in fact a Private function runned by the engine. You should
-  never use it without knowing very very well what you're doing.
+  * The arguments are explained just in case.
+  * @param $activityId is the activity that is being completed, when this is not
+  * passed the engine takes it from the $_REQUEST array,all activities
+  * are executed passing the activityId in the URI.
+  * @param $force indicates that the instance must be routed no matter if the
+  * activity is auto-routing or not. This is used when "sending" an
+  * instance from a non-auto-routed activity to the next activity.
+  * @private
+  * YOU MUST NOT CALL sendAutorouted() for non-interactive activities since
+  * the engine does automatically complete and send automatic activities after
+  * executing them.
+  * This function is in fact a Private function runned by the engine. You should
+  * never use it without knowing very very well what you're doing.
+  * @return false in case of any problem, true if nothing was done and an array if something 
+  * done, like walk on transition and execution of an activity (see sendTo comments) or if
+  * this activity was a split activity (in this case the array contains a row for each following activity)
   */
   function sendAutorouted($activityId,$force=false)
   {
     $type = $this->getOne("select `wf_type` from `".GALAXIA_TABLE_PREFIX."activities` where `wf_activity_id`=?",array((int)$activityId));    
-    //If the activity ending is autorouted then send to the activity
-    if ($type != 'end') 
+    //on a end activity we have nothing to do
+    if ($type == 'end')
     {
-      if (($force) || ($this->getOne("select `wf_is_autorouted` from `".GALAXIA_TABLE_PREFIX."activities` where `wf_activity_id`=?",array($activityId)) == 'y'))
+      return true;
+    }
+    //If the activity ending is not autorouted then we have nothing to do
+    if (!(($force) || ($this->getOne("select `wf_is_autorouted` from `".GALAXIA_TABLE_PREFIX."activities` where `wf_activity_id`=?",array($activityId)) == 'y')))
+    {
+      return true;
+    }
+    //If the activity ending is autorouted then send to the activity
+    // Now determine where to send the instance
+    $query = "select `wf_act_to_id` from `".GALAXIA_TABLE_PREFIX."transitions` where `wf_act_from_id`=?";
+    $result = $this->query($query,array((int)$activityId));
+    $candidates = Array();
+    while ($res = $result->fetchRow()) 
+    {
+      $candidates[] = $res['wf_act_to_id'];
+    }  
+    if($type == 'split') 
+    {
+      $erase_from = false;
+      $num_candidates = count($candidates);
+      $returned_data = Array();
+      $i = 1;
+      foreach ($candidates as $cand) 
       {
-        // Now determine where to send the instance
-        $query = "select `wf_act_to_id` from `".GALAXIA_TABLE_PREFIX."transitions` where `wf_act_from_id`=?";
-        $result = $this->query($query,array((int)$activityId));
-        $candidates = Array();
-        while ($res = $result->fetchRow()) 
-        {
-          $candidates[] = $res['wf_act_to_id'];
-        }  
-        if($type == 'split') 
-        {
-          $erase_from = false;
-          $num_candidates = count($candidates);
-          $i = 1;
-          foreach ($candidates as $cand) 
-          {
-			// only erase split activity in instance when all the activities comming from the split have been set up
-			if ($i == $num_candidates)
-                        { 
-                          $erase_from = true;
-                        }
-                        $this->sendTo($activityId,$cand,$erase_from);
-			$i++;
-          }
-        } 
-        elseif($type == 'switch') 
-        {
-          if (in_array($this->nextActivity,$candidates))
-          {
-            $this->sendTo((int)$activityId,(int)$this->nextActivity);
-          } 
-          else 
-          {
-            trigger_error(tra('Fatal error: nextActivity does not match any candidate in autorouting switch activity'),E_USER_WARNING);
-          }
-        } 
-        else 
-        {
-          if (count($candidates)>1) 
-          {
-            trigger_error(tra('Fatal error: non-deterministic decision for autorouting activity'),E_USER_WARNING);
-          }
-          else 
-          {
-            $this->sendTo((int)$activityId,(int)$candidates[0]);
-          }
+        // only erase split activity in instance when all the activities comming from the split have been set up
+        if ($i == $num_candidates)
+        { 
+          $erase_from = true;
         }
+        $returned_data[$i] = $this->sendTo($activityId,$cand,$erase_from);
+        $i++;
+      }
+      return $returned_data;
+    } 
+    elseif($type == 'switch') 
+    {
+      if (in_array($this->nextActivity,$candidates))
+      {
+        return $this->sendTo((int)$activityId,(int)$this->nextActivity);
+      } 
+      else 
+      {
+        $this->error[] = tra('Fatal error: nextActivity does not match any candidate in autorouting switch activity');
+        return false;
+        //trigger_error(tra('Fatal error: nextActivity does not match any candidate in autorouting switch activity'),E_USER_WARNING);
+      }
+    } 
+    else 
+    {
+      if (count($candidates)>1) 
+      {
+        $this->error[] = tra('Fatal error: non-deterministic decision for autorouting activity');
+        return false;
+        //trigger_error(tra('Fatal error: non-deterministic decision for autorouting activity'),E_USER_WARNING);
+      }
+      else 
+      {
+        return $this->sendTo((int)$activityId,(int)$candidates[0]);
       }
     }
   }
@@ -701,14 +721,16 @@ class Instance extends Base {
     //$this->setActivityStatus($activityId,'aborted');
 
     // terminate the instance with status 'aborted'
-    $this->terminate('aborted');
+    return $this->terminate('aborted');
   }
   
   /*!
-  Terminates the instance marking the instance and the process
-  as completed. This is the end of a process.
-  Normally you should not call this method since it is automatically
-  called when an end activity is completed.
+  * Terminates the instance marking the instance and the process
+  * as completed. This is the end of a process.
+  * Normally you should not call this method since it is automatically
+  * called when an end activity is completed.
+  * @param $status is the final status, 'completed' by default
+  * @return true if everything was ok, false else
   */
   function terminate($status = 'completed') {
     //Set the status of the instance to completed
@@ -719,13 +741,27 @@ class Instance extends Base {
     $this->query($query,array((int)$this->instanceId));
     $this->status = $status;
     $this->activities = Array();
+    return true;
   }
   
   
   /*!
-  Sends the instance from some activity to another activity.
-  You should not call this method unless you know very very well what
-  you are doing.
+  * Sends the instance from some activity to another activity. (walk on a transition)
+  * You should not call this method unless you know very very well what
+  * you are doing.
+  * @param $from is the activity id at the start of the transition
+  * @param $activityId is the activity id at the end of the transition
+  * @param $erase_from is true by default, if true the coming activity row will be erased from
+  * instance_activities table. You should set it to false for example with split activities while
+  * you still want to re-call this function
+  * @return false if anything goes wrong, true if we are at the end of the execution tree and an array
+  * if a part of the process was automatically runned at the end of the transition. this array contains
+  * 2 keys 'transition' is the transition we walked on, 'activity' is the result of the run part if it was an automatic activity.
+  * 'activity' value is an associated array containing several usefull keys:
+  *	* 'completed' is a boolean indicating that the activity was completed or not
+  *	* 'debug contains debug messages
+  *	* 'info' contains some usefull infos about the activity-instance running (like names)
+  *	* 'next' is the result of a SendAutorouted part which could in fact be the result of a call to this function, etc.
   */
   function sendTo($from,$activityId,$erase_from=true) {
     //1: if we are in a join check
@@ -736,7 +772,9 @@ class Instance extends Base {
     
     // Verify the existance of a transition
     if(!$this->getOne("select count(*) from `".GALAXIA_TABLE_PREFIX."transitions` where `wf_act_from_id`=? and `wf_act_to_id`=?",array($from,(int)$activityId))) {
-      trigger_error(tra('Fatal error: trying to send an instance to an activity but no transition found'),E_USER_WARNING);
+      $this->error[] = tra('Fatal error: trying to send an instance to an activity but no transition found');
+      return false;
+      //trigger_error(tra('Fatal error: trying to send an instance to an activity but no transition found'),E_USER_WARNING);
     }
 
     //init
@@ -812,11 +850,15 @@ class Instance extends Base {
         // which will let user having the good role mapping grab this activity later
         $putuser = $default_user;
       }
-    }        
+    }
+    //we will use an array for return value
+    $returned_data = Array();
+    
     //update the instance_activities table
     //if not splitting delete first
     //please update started,status,user
-    if($erase_from) {
+    if (($erase_from) && (!empty($this->instanceId)))
+    {
       $query = "delete from `".GALAXIA_TABLE_PREFIX."instance_activities` where `wf_instance_id`=? and `wf_activity_id`=?";
       $this->query($query,array((int)$this->instanceId,$from));
     }
@@ -824,16 +866,24 @@ class Instance extends Base {
     if ($type == 'join') {
       if (count($this->activities)>1) {
         // This instance will have to wait!
-        return;
+        $returned_data['transition']['status'] = 'waiting';
+        return $returned_data;
       }
     }    
 
+    //create the new instance-activity
     $now = date("U");
     $iid = $this->instanceId;
     $query="delete from `".GALAXIA_TABLE_PREFIX."instance_activities` where `wf_instance_id`=? and `wf_activity_id`=?";
     $this->query($query,array((int)$iid,(int)$activityId));
+$returned_data['transition']['debug'][]= $query. "::".$iid."::".$activityId;
     $query="insert into `".GALAXIA_TABLE_PREFIX."instance_activities`(`wf_instance_id`,`wf_activity_id`,`wf_user`,`wf_status`,`wf_started`) values(?,?,?,?,?)";
     $this->query($query,array((int)$iid,(int)$activityId,$putuser,'running',(int)$now));
+$returned_data['transition']['debug'][]= $query. "::".$iid."::".$$activityId."::".$putuser."::".'running et Now()';
+    
+    //record the transition walk
+    $returned_data['transition']['status'] = 'done';
+    $returned_data['transition']['target'] = $activityId;
     
     //we are now in a new activity
     $this->activities=Array();
@@ -847,16 +897,47 @@ class Instance extends Base {
     //execute the code for the activity and
     //complete the activity
     $isInteractive = $this->getOne("select `wf_is_interactive` from `".GALAXIA_TABLE_PREFIX."activities` where `wf_activity_id`=?",array((int)$activityId));
-    if ($isInteractive=='n') {
-
+    if ($isInteractive=='n') 
+    {
       // Now execute the code for the activity (function defined in galaxia's config.php)
-      galaxia_execute_activity($activityId, $iid , 1);
-
+      $returned_data['activity'] =& galaxia_execute_activity($activityId, $iid , 1);
+      //we should have some info in $returned_data now. if it is false there's a problem
+      if ((!(is_array($returned_data['activity']))) && (!($returned_data['activity'])) )
+      {
+        $this->error[] = tra('failed to execute automatic activity');
+        //record the failure
+        $returned_data['activity']['failure'] = true;
+        return $returned_data;
+      }
+      else
+      {
+        //ok, we have an array, but it can still be a bad result
+        //this one is just for debug info
+        if (isset($returned_data['activity']['debug']))
+        {
+          //we retrieve this info here, in this object
+          $this->error[] = $returned_data['activity']['debug'];
+        }
+        //and this really test if it worked, if not we have a nice failure message (better than just failure=true)
+        if (isset($returned_data['activity']['failure']))
+        {
+          $this->error[] = tra('failed to execute automatic activity');
+          $this->error[] = $returned_data['activity']['failure'];
+          //record the failure
+          return $returned_data;
+        }
+      }
       // Reload in case the activity did some change
       $this->getInstance($this->instanceId);
-      $this->complete($activityId);
-      $this->sendAutorouted($activityId);
+      //complete the activity
+      if ($this->complete($activityId))
+      {
+        $returned_data['activity']['completed']=true;
+      }
+      //and send the next autorouted activity if any
+      $returned_data['activity']['next'] = $this->sendAutorouted($activityId);
     }
+    return $returned_data;
   }
   
   /*!
